+++ test/unit-tests/json/replacer.test.js
@@ -1,11 +1,20 @@
 import assert from 'assert'
 import math from '../../../src/bundleAny'
+const replacer = math.replacer
 
 describe('replacer', function () {
   it('should stringify generic JSON', function () {
     const data = { foo: [1, 2, 3], bar: null, baz: 'str' }
     const json = '{"foo":[1,2,3],"bar":null,"baz":"str"}'
     assert.deepStrictEqual(JSON.stringify(data), json)
+    assert.deepStrictEqual(JSON.stringify(data, replacer), json)
+  })
+
+  it('should stringify a number with special values like Infinity', function () {
+    assert.deepStrictEqual(JSON.stringify(2.3, replacer), '2.3')
+    assert.deepStrictEqual(JSON.stringify(Infinity, replacer), '{"mathjs":"number","value":"Infinity"}')
+    assert.deepStrictEqual(JSON.stringify(-Infinity, replacer), '{"mathjs":"number","value":"-Infinity"}')
+    assert.deepStrictEqual(JSON.stringify(NaN, replacer), '{"mathjs":"number","value":"NaN"}')
   })
 
   it('should stringify a Complex number', function () {
@@ -13,6 +22,7 @@ describe('replacer', function () {
     const json = '{"mathjs":"Complex","re":2,"im":4}'
 
     assert.deepStrictEqual(JSON.stringify(c), json)
+    assert.deepStrictEqual(JSON.stringify(c, replacer), json)
   })
 
   it('should stringify a BigNumber', function () {
@@ -20,6 +30,7 @@ describe('replacer', function () {
     const json = '{"mathjs":"BigNumber","value":"5"}'
 
     assert.deepStrictEqual(JSON.stringify(b), json)
+    assert.deepStrictEqual(JSON.stringify(b, replacer), json)
   })
 
   it('should stringify a Fraction', function () {
@@ -27,12 +38,14 @@ describe('replacer', function () {
     const json = '{"mathjs":"Fraction","n":3,"d":8}'
 
     assert.deepStrictEqual(JSON.stringify(b), json)
+    assert.deepStrictEqual(JSON.stringify(b, replacer), json)
   })
 
   it('should stringify a Range', function () {
     const r = new math.Range(2, 10)
     const json = '{"mathjs":"Range","start":2,"end":10,"step":1}'
     assert.deepStrictEqual(JSON.stringify(r), json)
+    assert.deepStrictEqual(JSON.stringify(r, replacer), json)
   })
 
   it('should stringify an Index', function () {
@@ -42,18 +55,21 @@ describe('replacer', function () {
         '{"mathjs":"ImmutableDenseMatrix","data":[2],"size":[1]}' +
         ']}'
     assert.deepStrictEqual(JSON.stringify(i), json)
+    assert.deepStrictEqual(JSON.stringify(i, replacer), json)
   })
 
   it('should stringify a Range (2)', function () {
     const r = new math.Range(2, 10, 2)
     const json = '{"mathjs":"Range","start":2,"end":10,"step":2}'
     assert.deepStrictEqual(JSON.stringify(r), json)
+    assert.deepStrictEqual(JSON.stringify(r, replacer), json)
   })
 
   it('should stringify a Unit', function () {
     const u = new math.Unit(5, 'cm')
     const json = '{"mathjs":"Unit","value":5,"unit":"cm","fixPrefix":false}'
     assert.deepStrictEqual(JSON.stringify(u), json)
+    assert.deepStrictEqual(JSON.stringify(u, replacer), json)
   })
 
   it('should stringify a Matrix, dense', function () {
@@ -61,6 +77,7 @@ describe('replacer', function () {
     const json = '{"mathjs":"DenseMatrix","data":[[1,2],[3,4]],"size":[2,2]}'
 
     assert.deepStrictEqual(JSON.stringify(m), json)
+    assert.deepStrictEqual(JSON.stringify(m, replacer), json)
   })
 
   it('should stringify a Matrix, sparse', function () {
@@ -68,12 +85,14 @@ describe('replacer', function () {
     const json = '{"mathjs":"SparseMatrix","values":[1,3,2,4],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}'
 
     assert.deepStrictEqual(JSON.stringify(m), json)
+    assert.deepStrictEqual(JSON.stringify(m, replacer), json)
   })
 
   it('should stringify a ResultSet', function () {
     const r = new math.ResultSet([1, 2, new math.Complex(3, 4)])
     const json = '{"mathjs":"ResultSet","entries":[1,2,{"mathjs":"Complex","re":3,"im":4}]}'
     assert.deepStrictEqual(JSON.stringify(r), json)
+    assert.deepStrictEqual(JSON.stringify(r, replacer), json)
   })
 
   it('should stringify a Matrix containing a complex number, dense', function () {
@@ -82,6 +101,7 @@ describe('replacer', function () {
     const json = '{"mathjs":"DenseMatrix","data":[[1,2],[3,{"mathjs":"Complex","re":4,"im":5}]],"size":[2,2]}'
 
     assert.deepStrictEqual(JSON.stringify(m), json)
+    assert.deepStrictEqual(JSON.stringify(m, replacer), json)
   })
 
   it('should stringify a Matrix containing a complex number, sparse', function () {
@@ -90,12 +110,14 @@ describe('replacer', function () {
     const json = '{"mathjs":"SparseMatrix","values":[1,3,2,{"mathjs":"Complex","re":4,"im":5}],"index":[0,1,0,1],"ptr":[0,2,4],"size":[2,2]}'
 
     assert.deepStrictEqual(JSON.stringify(m), json)
+    assert.deepStrictEqual(JSON.stringify(m, replacer), json)
   })
 
   it('should stringify a Chain', function () {
     const c = math.chain(2.3)
     const json = '{"mathjs":"Chain","value":2.3}'
     assert.deepStrictEqual(JSON.stringify(c), json)
+    assert.deepStrictEqual(JSON.stringify(c, replacer), json)
   })
 
   it('should stringify a node tree', function () {
@@ -139,11 +161,13 @@ describe('replacer', function () {
     }
 
     assert.deepStrictEqual(JSON.parse(JSON.stringify(node)), json)
+    assert.deepStrictEqual(JSON.parse(JSON.stringify(node, replacer)), json)
   })
 
   it('should stringify Help', function () {
     const h = new math.Help({ name: 'foo', description: 'bar' })
     const json = '{"mathjs":"Help","name":"foo","description":"bar"}'
     assert.deepStrictEqual(JSON.parse(JSON.stringify(h)), JSON.parse(json))
+    assert.deepStrictEqual(JSON.parse(JSON.stringify(h, replacer)), JSON.parse(json))
   })
 })